fixup! promote i64 amounts to i128

+++ masp_proofs/src/sapling/verifier.rs
@@ -168,7 +168,7 @@ impl SaplingVerificationContext {
     /// have been checked before calling this function.
     pub fn final_check(
         &self,
-        assets_and_values: &[(AssetType, i64)],
+        assets_and_values: &[(AssetType, i128)],
         sighash_value: &[u8; 32],
         binding_sig: Signature,
     ) -> bool {
